relocate scan and search button
* relocated scan and search button based on user-testing feedback
* moved my_pi to home bottom navigation
